+++ churn/src/main/java/com/challenge/churn/controller/EmpresaController.java
@@ -3,6 +3,8 @@ package com.challenge.churn.controller;
 import com.challenge.churn.model.Empresa;
 import com.challenge.churn.repo.EmpresaRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,25 +21,42 @@ public class EmpresaController {
     }
 
     @GetMapping(value = "/empresa/{id}")
-    public Empresa getEmpresas(@PathVariable long id){
-        return empresaRepo.findById(id).get();
+    public ResponseEntity<?> getEmpresa(@PathVariable long id){
+        Empresa empresa = empresaRepo.findById(id).orElse(null);
+        if (empresa != null) {
+            return ResponseEntity.ok(empresa);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping(value = "/empresa/create")
-    public void saveEmpresa(@RequestBody Empresa empresa){
+    public ResponseEntity<?> saveEmpresa(@RequestBody Empresa empresa){
         empresaRepo.save(empresa);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping(value = "/empresa/update/{id}")
-    public void updateEmpresa(@PathVariable long id, @RequestBody Empresa empresa){
-        Empresa updatedEmpresa = empresaRepo.findById(id).get();
-        updatedEmpresa.setNome(empresa.getNome());
-        updatedEmpresa.setEndereco(empresa.getEndereco());
+    public ResponseEntity<?> updateEmpresa(@PathVariable long id, @RequestBody Empresa empresa){
+        Empresa updatedEmpresa = empresaRepo.findById(id).orElse(null);
+        if (updatedEmpresa != null) {
+            updatedEmpresa.setNome(empresa.getNome());
+            updatedEmpresa.setEndereco(empresa.getEndereco());
+            empresaRepo.save(updatedEmpresa);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
-    @DeleteMapping(value = "/delete/{id}")
-    public void deleteEmpresa(@PathVariable long id){
-        Empresa deletedEmpresa = empresaRepo.findById(id).get();
-        empresaRepo.delete(deletedEmpresa);
+    @DeleteMapping(value = "/empresa/delete/{id}")
+    public ResponseEntity<?> deleteEmpresa(@PathVariable long id){
+        Empresa deletedEmpresa = empresaRepo.findById(id).orElse(null);
+        if (deletedEmpresa != null) {
+            empresaRepo.delete(deletedEmpresa);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
