+++ churn/src/main/java/com/challenge/churn/controller/TransacaoController.java
@@ -3,6 +3,8 @@ package com.challenge.churn.controller;
 import com.challenge.churn.model.Transacao;
 import com.challenge.churn.repo.TransacaoRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,31 +16,49 @@ public class TransacaoController {
     private TransacaoRepo transacaoRepo;
 
     @GetMapping(value = "/transacao")
-    public List<Transacao> getTransacoes(){
-        return transacaoRepo.findAll();
+    public ResponseEntity<List<Transacao>> getTransacoes(){
+        List<Transacao> transacoes = transacaoRepo.findAll();
+        return ResponseEntity.ok(transacoes);
     }
 
     @GetMapping(value = "/transacao/{id}")
-    public Transacao getTransacoes(@PathVariable long id){
-        return transacaoRepo.findById(id).get();
+    public ResponseEntity<?> getTransacao(@PathVariable long id){
+        Transacao transacao = transacaoRepo.findById(id).orElse(null);
+        if (transacao != null) {
+            return ResponseEntity.ok(transacao);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping(value = "/transacao/create")
-    public void saveTransacao(@RequestBody Transacao transacao){
+    public ResponseEntity<?> saveTransacao(@RequestBody Transacao transacao){
         transacaoRepo.save(transacao);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @PutMapping(value = "/transacao/update/{id}")
-    public void updateTransacao(@PathVariable long id, @RequestBody Transacao transacao){
-        Transacao updatedTransacao = transacaoRepo.findById(id).get();
-        updatedTransacao.setCliente(transacao.getCliente());
-        updatedTransacao.setValor(transacao.getValor());
-        updatedTransacao.setData(transacao.getData());
+    public ResponseEntity<?> updateTransacao(@PathVariable long id, @RequestBody Transacao transacao){
+        Transacao updatedTransacao = transacaoRepo.findById(id).orElse(null);
+        if (updatedTransacao != null) {
+            updatedTransacao.setCliente(transacao.getCliente());
+            updatedTransacao.setValor(transacao.getValor());
+            updatedTransacao.setData(transacao.getData());
+            transacaoRepo.save(updatedTransacao);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
-    @DeleteMapping(value = "/delete/{id}")
-    public void deleteTransacao(@PathVariable long id){
-        Transacao deletedTransacao = transacaoRepo.findById(id).get();
-        transacaoRepo.delete(deletedTransacao);
+    @DeleteMapping(value = "/transacao/delete/{id}")
+    public ResponseEntity<?> deleteTransacao(@PathVariable long id){
+        Transacao deletedTransacao = transacaoRepo.findById(id).orElse(null);
+        if (deletedTransacao != null) {
+            transacaoRepo.delete(deletedTransacao);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
